requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-filter==22.1
django-guardian==2.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
Pygments==2.13.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
